refactor(add tests): move serializers from views to new folder and add tests
