+++ backend/api/serializers/user.py
@@ -1,0 +1,21 @@
+"""User DRF serializers file"""
+
+from django.contrib.auth.models import User
+
+from rest_framework import serializers
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ['id', 'username', 'email']
+
+class UserCreateSerializer(serializers.Serializer):
+    username = serializers.CharField()
+    email = serializers.EmailField()
+    password = serializers.CharField()
+
+    def create(self, validated_data):
+        return User.objects.create_user(
+            validated_data['username'], validated_data['email'], validated_data['password']
+        )
