+++ backend/api/views/annotation_set.py
@@ -1,23 +1,10 @@
 """Annotation set DRF-Viewset file"""
 
-from rest_framework import viewsets, serializers
+from rest_framework import viewsets
 from rest_framework.response import Response
-from rest_framework.decorators import action
-
-from drf_spectacular.utils import extend_schema_field
 
 from backend.api.models import AnnotationSet
-
-class AnnotationSetSerializer(serializers.ModelSerializer):
-    tags = serializers.SerializerMethodField()
-
-    class Meta:
-        model = AnnotationSet
-        fields = ['id', 'name', 'desc', 'tags']
-
-    @extend_schema_field(serializers.ListField(child=serializers.CharField()))
-    def get_tags(self, annotation_set):
-        return list(annotation_set.tags.values_list('name', flat=True))
+from backend.api.serializers import AnnotationSetSerializer
 
 class AnnotationSetViewSet(viewsets.ViewSet):
     """
