+++ backend/tests/__init__.py
@@ -3,3 +3,5 @@ from .views.dataset import *
 from .views.annotation_set import *
 from .views.annotation_task import *
 from .views.annotation_campaign import *
+from .serializers.annotation_task import *
+from .serializers.annotation_campaign import *
