+++ backend/tests/fixtures/annotation_campaigns_tasks.yaml
@@ -13,7 +13,6 @@
     - 1
     spectro_configs:
     - 1
-    - 2
 - model: api.annotationtask
   pk: 1
   fields:
